// src/infrastructure/database/schemas/taco-item.schema.ts
export interface TacoItemDocument {
  _id: string;
  codigo: string;
  nome: string;
  nomeCientifico?: string;
  grupo: string;
  marca?: string;
  energiaKcal: number;
  energiaKj: number;
  umidade: number;
  proteinaG: number;
  lipidioG: number;
  colesterolMg: number;
  carboidratoG: number;
  fibraAlimentarG: number;
  cinzasG: number;
  calcioMg: number;
  magnesioMg: number;
  manganesMg: number;
  fosforoMg: number;
  ferroMg: number;
  sodioMg: number;
  potassioMg: number;
  cobreMg: number;
  zincoMg: number;
  retinolMcg: number;
  reMcg: number;
  raeMcg: number;
  tiaminaMg: number;
  riboflavinaMg: number;
  piridoxinaMg: number;
  niacinaMg: number;
  vitaminaCMg: number;
  porcaoPadraoG: number;
  unidade: "g" | "ml";
  createdAt: Date;
  updatedAt: Date;
  searchableText: string;
}